Implement spilling of kernel SLB entries.  Mostly from FreeBSD.
This makes it possible to use more SLB entries for the kernel than the
hardware supports.  The design is such that a subset of the hardware SLB
entries can be replaced when needed.  This makes sure the entries
mapping kernel code and data and the page tables ar always present.
Traps for missing SLB entries are handled in real-mode and on a special
stack such that it doesn't have to rely on SLB entires mapping kernel
stacks.

With this in place we can increase KVA to 32GB.  Hopefully that's enough
to support large memory configurations.