similar to fix by kettenis on amd64:
The uvm_map_inentry() check may sleep to grab the lock of the map.
The fault address is read from cr2 in pageflttrap() which
gets called after this check and if the check sleeps, cr2 is likely to
be clobbered by a page fault in another process.

Fix this by reading cr2 early and pass it to pageflttrap().

ok kettenis