timeout(9): timeout_run(): read to_process before leaving timeout_mutex
to_process is assigned during timeout_add(9) within timeout_mutex.  In
timeout_run() we need to read to_process before leaving timeout_mutex
to ensure that the process pointer given to kcov_remote_enter(9) is
the same as the one we set from timeout_add(9) when the candidate
timeout was originally scheduled to run.