Document that `a_p' is always curproc by using a KASSERT().
One exception of this rule is VOP_CLOSE() where NULL is used instead
of curproc when the garbace collector of unix sockets, that runs in
a kernel thread, drops the last reference of a file.

This will allows for future simplifications of the VFS interfaces.

Previous version ok visa@, anton@.

ok kn@