apply the "depend on" logic in the the default handling of rtr links.
this fixes the use of "depend on" in my situation, which is an
Ethernet (broadcast) interface where I'm actually peering with
multiple routers so i don't have "type p2p" set. with this the ospf
peers now know how to route around my firewalls when their carp
interfaces are in the backup state.

discussed with claudio@ and jmatthew@
ok remi@