Fix handling of MSG_PEEK in soreceive() for the case where an empty mbuf is encountered in a seqpacket socket.
This diff uses the fact that setting orig_resid to 0 results in soreceive()
to return instead of looping back with the intent to sleep for more data.
orig_resid is now always set to 0 in the control message case (instead of
only if controlp is defined). This is the same behaviour as for the PR_NAME
case.  Additionally orig_resid is set to 0 in the data reader when MSG_PEEK
is used.

Tested in snaps for a while and by anton@

Reported-by: syzbot+4b0e9698b344b0028b14@syzkaller.appspotmail.com