hvn(4): msleep(9) -> msleep_nsec(9)
In hvn_alloc_cmd() we may spin waiting for a free rndis_cmd.  We check
the list once per tick and block with msleep(9) if there aren't any
free objects.

In practice though we don't need to poll for a free rndis_cmd because
our sleep is protected by a mutex, so we can't miss a wakeup(9).  That
is, it's safe to use msleep_nsec(9) here and not set a timeout (INFSLP).

Tested by Andre Stoebe <as@nul.space> (Hyper-V on Windows 10).

"LGTM" mikeb@, ok mpi@