an(4): tsleep(9) -> tsleep_nsec(9)
In an_wait() we spin for up to (3 * hz) iterations awaiting AN_EV_CMD.
If we don't see it in a given iteration we block for up to 1 tick and
spin again.

jsg@ suggests instead blocking for 100ms at a time for up to 30
iterations.

Discussed with mpi@.  Idea from jsg@.  Murmurs of agreement from
claudio@ and kettenis@.

probably ok jsg@