go to splhigh around the kernel clock and hardware timer reads.
the idea is to avoid some other work, like a hardware interrupt,
running in between the reads of the kernel and chip clocks and
therefore skewing the interval calculations. this tightens up a
lot of the slop seen when using the cqe timestamp for an mbuf
timestamp, but there's still something not quite right.