stop showing amd l3 cache information
>= zen 2 based amd processors return a value of 9 for l3 cache assoc
via cpuid 0x80000006.  As that is a reserved value we end up incorrectly
claiming the l3 cache is disabled.  While it is possible to get l3 cache
information via cpuid 0x8000001d when TOPEXT is advertised that will
instead give information about the l3 cache available to the core
complex (CCX) that the cpu belongs to where previously the amount of l3
available to all core complexes was shown.

As we don't detail topology in dmesg or show the mapping of cores to
core complexes just stop displaying l3 information.  It already isn't
shown on intel.

ok gkoehler@