Fix boot-time crash on sparc64
On sparc64, initmsgbuf() is invoked before curcpu() is usable
on the boot processor. Consequently, it is unsafe to use mutexes
during the message buffer initialization. Avoid such use by skipping
log_mtx when appending a newline from initmsgbuf().

Use mbp instead of msgbufp as the buffer argument to the putchar routine
for consistency.

Bug reported and fix suggested by miod@