remove the suggestion to permit pkg_add with doas "nopass" when doing ports dev work.
if you are able to run pkg_add as root without a password, your account
is root-equivalent.

typing the password multiple times is a pain but if somebody is going to
choose to weaken their local security in this way, it should be their
own decision and not something they have read in a manpage.

ok tb@ thfr@